pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bakalari-mcp" 
version = "1.0.0"
description = "MCP Server pro Bakaláři v3 API"
authors = [
    {name = "Miroslav Dvořák", email = "mirecekd@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "fastmcp>=0.2.0",
    "aiohttp>=3.8.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
bakalari-mcp-server = "bakalari_mcp_server.server:main"

[project.urls]
Homepage = "https://github.com/mirecekd/bakalari-mcp"
Repository = "https://github.com/mirecekd/bakalari-mcp"
Issues = "https://github.com/mirecekd/bakalari-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/bakalari_mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
    "/LICENSE",
]
